CORS bug fixes
- simplified error handling.
- added code for a timeout.

I faced issues while running Cors scan on a list of URLs and there was a URL in the list with some connection problem. Due to this, the script got exited. So I improved the error handling and now it is working fine and also showing the URLs with connection problems.